refactor: enum type 패키지 이동
