+++ src/main/java/sisosolsol/greenfire/common/enums/post/PostType.java
@@ -1,4 +1,4 @@
-package sisosolsol.greenfire.post.model.type;
+package sisosolsol.greenfire.common.enums.post;
 
 public enum PostType {
     DEFAULT,
